Refactor import paths for comment-related components and remove unused imports
- Refactored import paths for `NewCommentButton` and `AuthCommentButton` components in the comments section
- Removed unused imports in the navbar component
- Removed unused imports and refactored import path for `Badge` component in the popular entries section
