+++ src/components/public/comments/auth-comment-button.tsx
@@ -1,7 +1,7 @@
 import { Plate } from '@/lib/plates';
 import { signIn, auth } from '@/auth';
 import { Button } from '@/components/ui/button';
-import NewCommentButton from '@/components/comments/new-comment-button';
+import NewCommentButton from '@/components/public/comments/new-comment-button';
 
 interface AuthCommentButtonProps {
   className?: string;
