+++ src/components/public/navbar.tsx
@@ -1,8 +1,6 @@
 import Link from 'next/link';
 import React from 'react';
 import { ModeToggle } from '@/components/ui/mode-toggle';
-import { signIn, signOut } from '@/auth';
-import { Button } from '@/components/ui/button';
 import AuthButton from '@/components/public/auth-button';
 
 const Navbar = () => {
